+++ PointAnalysis/meshmodel.h
@@ -45,8 +45,8 @@ public:
 	int faceCount();
 	Eigen::Vector3f getCentroid() const;
 	double getRadius() const { return m_radius; }
-	void draw(int scale);
-	void drawSamples(int scale);
+	void draw(float scale);
+	void drawSamples(float scale);
 	void samplePoints();
 	std::string getInputFilepath() const;
 	QVector<int> getVerticesLabels() const;
